register page partial button to component
