+++ src/modules/auth/pages/register/template.hbs
@@ -1,33 +1,35 @@
 <div class="container">
-    <div class="auth-wrapper">
-        <div class="card bordered bg-success accent">
-            <form id="register-form">
+  <div class="auth-wrapper">
+    <div class="card bordered bg-success accent">
+      <form id="register-form">
 
-                <h3 class="card__title">Регистрация</h3>
+          <h3 class="card__title">Регистрация</h3>
 
-                {{> inputText input=form.email }}
-                {{> inputText input=form.login }}
-                {{> inputText input=form.first_name }}
-                {{> inputText input=form.second_name }}
-                {{> inputText input=form.phone }}
-                {{> inputText input=form.password }}
-                {{> inputText input=form.passwordConfirm }}
+          {{> inputText input=form.email }}
+          {{> inputText input=form.login }}
+          {{> inputText input=form.first_name }}
+          {{> inputText input=form.second_name }}
+          {{> inputText input=form.phone }}
+          {{> inputText input=form.password }}
+          {{> inputText input=form.passwordConfirm }}
 
-            </form>
+      </form>
 
-            <div class="spacer"></div>
+      <div class="spacer"></div>
 
-            {{> button id=registerBtnId value='Создать аккаунт' class='bg-primary white' }}
-            <img
-                    id="{{ loadingId }}"
-                    class="loading"
-                    style="display: none"
-                    src="images/loading.gif"
-                    alt="loading"
-                    height="40px"
-            >
-            {{> button id=goHomeBtnId href='/' value='Уже зарегистрированы?' outline=true class="white" }}
+      {{{ RegisterBtn }}}
 
-        </div>
+      <img
+              id="{{ loadingId }}"
+              class="loading"
+              style="display: none"
+              src="images/loading.gif"
+              alt="loading"
+              height="40px"
+      >
+
+      {{{ RedirectLoginBtn }}}
+
     </div>
+  </div>
 </div>
